# Speed up ZODB.blob.BushyLayout.oid_to_path on Python 3